// src/pages/Subject/form/template/fields/types.ts
export enum FIELD_TYPES {
  website_url = "website_url",
  email_address = "email_address",
  images = "images",
  title = "title",
  social_media = "social_media",
  phones = "phones",
  top_block_divider = "top_block_divider",
  bottom_block_divider = "bottom_block_divider",
  block_divider = "block_divider",
  empty_divider = "empty_divider",
}